+++ frontends/web/src/components/amount/amount-with-unit.tsx
@@ -1,7 +1,7 @@
 // SPDX-License-Identifier: Apache-2.0
 
 import { useContext } from 'react';
-import type { CoinUnit, ConversionUnit, TAmountWithConversions } from '@/api/account';
+import type { NativeCoinUnit, ConversionUnit, TAmountWithConversions } from '@/api/account';
 import { RatesContext } from '@/contexts/RatesContext';
 import { Amount } from '@/components/amount/amount';
 import { isBitcoinCoin } from '@/routes/account/utils';
@@ -13,7 +13,9 @@ type TAmountWithUnitProps = {
   sign?: string;
   alwaysShowAmounts?: boolean;
   convertToFiat?: boolean;
+  amountClassName?: string;
   unitClassName?: string;
+  maxDecimals?: number;
 };
 
 export const AmountWithUnit = ({
@@ -22,7 +24,9 @@ export const AmountWithUnit = ({
   sign,
   convertToFiat,
   alwaysShowAmounts = false,
+  amountClassName = '',
   unitClassName = '',
+  maxDecimals,
 }: TAmountWithUnitProps) => {
   const { rotateDefaultCurrency, defaultCurrency, rotateBtcUnit } = useContext(RatesContext);
 
@@ -30,7 +34,7 @@ export const AmountWithUnit = ({
     return null;
   }
   let displayedAmount: string = '';
-  let displayedUnit: CoinUnit | ConversionUnit;
+  let displayedUnit: NativeCoinUnit | ConversionUnit;
   let onClick: () => Promise<void>;
 
   if (convertToFiat) {
@@ -53,6 +57,7 @@ export const AmountWithUnit = ({
       ${style.rates || ''}
       ${style.availableFiatAmount || ''}
       ${!displayedAmount && style.notAvailable || ''}
+      ${amountClassName || ''}
     `.trim()}>
       {!!displayedAmount ? sign : ''}
       {!!displayedAmount ? (
@@ -61,6 +66,7 @@ export const AmountWithUnit = ({
           amount={displayedAmount}
           unit={displayedUnit}
           onMobileClick={enableClick ? onClick : undefined}
+          maxDecimals={maxDecimals}
         />
       ) : '---'}
       {' '}
@@ -74,7 +80,7 @@ export const AmountWithUnit = ({
 
 type TAmountUnitProps = {
   rotateUnit?: () => Promise<void>;
-  unit: ConversionUnit | CoinUnit;
+  unit: ConversionUnit | NativeCoinUnit;
   className?: string;
 };