+++ frontends/web/src/routes/account/info/signingconfiguration.tsx
@@ -1,14 +1,19 @@
 // SPDX-License-Identifier: Apache-2.0
 
-import { useState, ReactNode } from 'react';
+import { useState, useEffect, ReactNode } from 'react';
 import { useNavigate } from 'react-router-dom';
 import { useTranslation } from 'react-i18next';
-import { AccountCode, TAccount, TBitcoinSimple, TEthereumSimple, TSigningConfiguration, verifyXPub } from '@/api/account';
-import { getScriptName, isBitcoinBased } from '@/routes/account/utils';
+import { AccountCode, TAccount, TAmountWithConversions, TBitcoinSimple, TEthereumSimple, TSigningConfiguration, verifyXPub, getBalance } from '@/api/account';
+import { getScriptName, isBitcoinBased, isEthereumBased } from '@/routes/account/utils';
 import { alertUser } from '@/components/alert/Alert';
+import { AmountWithUnit } from '@/components/amount/amount-with-unit';
+import { A } from '@/components/anchor/anchor';
 import { CopyableInput } from '@/components/copy/Copy';
 import { Button } from '@/components/forms';
+import { ExternalLink } from '@/components/icon';
+import { Message } from '@/components/message/message';
 import { QRCode } from '@/components/qrcode/qrcode';
+import { truncateMiddle } from '@/utils/address';
 import style from './info.module.css';
 
 type TProps = {
@@ -23,6 +28,17 @@ export const SigningConfiguration = ({ account, info, code, signingConfigIndex, 
   const navigate = useNavigate();
   const { t } = useTranslation();
   const [verifying, setVerifying] = useState(false);
+  const [balance, setBalance] = useState<TAmountWithConversions>();
+
+  useEffect(() => {
+    if (isEthereumBased(account.coinCode)) {
+      getBalance(code).then(response => {
+        if (response.success) {
+          setBalance(response.balance.available);
+        }
+      }).catch(console.error);
+    }
+  }, [code, account.coinCode]);
 
   const getSimpleInfo = (): TBitcoinSimple | TEthereumSimple => {
     if (info.bitcoinSimple !== undefined) {
@@ -33,12 +49,21 @@ export const SigningConfiguration = ({ account, info, code, signingConfigIndex, 
 
   const config = getSimpleInfo();
   const bitcoinBased = isBitcoinBased(account.coinCode);
+  const contractAddress = account.contractAddress;
+  const blockExplorerAddressPrefix = account.blockExplorerAddressPrefix;
+  const contractAddressInfo = account.isToken && contractAddress && blockExplorerAddressPrefix
+    ? {
+      address: contractAddress,
+      url: `${blockExplorerAddressPrefix}${contractAddress}`,
+    }
+    : null;
   return (
     <div className={style.address}>
       <div className={style.qrCode}>
         { bitcoinBased ? (
           <QRCode
-            data={config.keyInfo.xpub} />
+            data={config.keyInfo.xpub}
+            size={220} />
         ) : null }
       </div>
       <div className={style.details}>
@@ -67,19 +92,44 @@ export const SigningConfiguration = ({ account, info, code, signingConfigIndex, 
           <strong>{account.isToken ? 'Token' : 'Coin'}:</strong>
           <span>{account.coinName} ({account.coinUnit})</span>
         </div>
+        { balance ? (
+          <div key="balance" className={style.entry}>
+            <strong>{t('accountSummary.balance')}:</strong>
+            <span>
+              <AmountWithUnit amount={balance} />
+            </span>
+          </div>
+        ) : null }
+        { contractAddressInfo ? (
+          <div key="contractAddress" className={style.entry}>
+            <strong>{t('accountInfo.contractAddress')}:</strong>
+            <div className={style.contractAddressLink}>
+              <code>{truncateMiddle(contractAddressInfo.address)}</code>
+              <A
+                href={contractAddressInfo.url}
+                title={contractAddressInfo.url}>
+                <ExternalLink className={style.contractAddressLinkIcon} />
+              </A>
+            </div>
+          </div>
+        ) : null}
         { bitcoinBased ? (
           <div key="xpub" className={`${style.entry || ''} ${style.largeEntry || ''}`}>
             <strong className="m-right-half">
               {t('accountInfo.extendedPublicKey')}:
             </strong>
             <CopyableInput
-              className="flex-grow"
               alignLeft
               flexibleHeight
               value={config.keyInfo.xpub} />
           </div>
         ) : null }
       </div>
+      { contractAddressInfo ? (
+        <Message className={style.warningMessage} type="warning">
+          {t('accountInfo.contractAddressWarning')}
+        </Message>
+      ) : null}
       <div className={style.buttons}>
         { bitcoinBased ? (
           <Button className={style.verifyButton} primary disabled={verifying} onClick={async () => {